fixed the api and setting it ready for deploy

+++ app/controllers/api/v1/books_controller.rb
@@ -1,9 +1,10 @@
 class Api::V1::BooksController < ApplicationController
   before_action :authenticate_user!
+  load_and_authorize_resource
 
   def index
-    books = Book.includes(:images)
-    render json: books, include: %i[images]
+    books = Book.includes(:image)
+    render json: books, include: %i[image]
   end
 
   def show
@@ -12,10 +13,11 @@ class Api::V1::BooksController < ApplicationController
   end
 
   def create
-    book = Book.new(book_params)
+    book = Book.new(book_params.except(:image_file))
+
     if book.save
-      attach_image(book)
-      render json: book, status: :created
+      attach_image(book) if params[:book][:image_file].present?
+      render json: { book:, status: :created, message: 'Book created successfully' }
     else
       render json: book.errors, status: :unprocessable_entity
     end
@@ -23,9 +25,10 @@ class Api::V1::BooksController < ApplicationController
 
   def update
     book = Book.find(params[:id])
-    if book.update(book_params)
-      attach_image(book)
-      render json: book, status: :ok
+    attach_image(book) if params[:image_file].present?
+
+    if book.update(book_params.except(:image_file))
+      render json: { book:, status: :ok, message: 'Book updated successfully' }
     else
       render json: book.errors, status: :unprocessable_entity
     end
@@ -33,14 +36,30 @@ class Api::V1::BooksController < ApplicationController
 
   def destroy
     book = Book.find(params[:id])
-    book.destroy
-    head :no_content
+
+    if book.reserved? || book.borrowed?
+      render json: { status: :unprocessable_entity,
+                     message: 'Book cannot be deleted because it is reserved or borrowed' }
+    else
+      book.destroy
+      render json: { status: :success, message: 'Book deleted successfully' }
+    end
   end
 
   private
 
+  def attach_image(book)
+    image_url = params[:book][:image_file]
+    return unless image_url.present?
+
+    book.build_image(image_data: image_url)
+    return if book.image.save
+
+    render json: book.image.errors, status: :unprocessable_entity and return
+  end
+
   def book_params
-    params.require(:book).permit(:title, :author, :description, :category_id, :recommended, :file_url,
-                                 image_attributes: %i[id image_file])
+    params.require(:book).permit(:title, :author, :description, :file_url, :category_id, :recommended,
+                                 :image_file)
   end
 end
